imageAPIのdelete処理でエラーハンドリングおよびmulterの責務を修正した

+++ api/src/usecases/image.js
@@ -1,33 +1,45 @@
 import * as imageRepository from "../repositories/image.js";
+import Result from "../utils/commons/Result.js";
+import AppError from "../utils/commons/AppError.js";
 
-export async function createImage({ fileInfo }) {
-  const { tempPath } = fileInfo;
-  const imageId = await imageRepository.insertRecord();
-  const uploadFileName = imageRepository.nameFile(imageId, tempPath);
-  const result = await imageRepository.moveFile(tempPath, uploadFileName);
-  return result;
+export async function createImage(tempPath) {
+  const insertResult = await imageRepository.insertRecord();
+  if (insertResult.isFailure) return insertResult;
+  const imageId = insertResult.data;
+
+  const nameResult = imageRepository.nameFile(imageId, tempPath);
+  if (nameResult.isFailure) return nameResult;
+  const uploadFileName = nameResult.data;
+
+  const moveResult = await imageRepository.moveFile(tempPath, uploadFileName);
+  if (moveResult.isFailure) return moveResult;
+
+  return insertResult;
 }
 
 //画像の取得
-export async function getImageById(imageId) {
-  const result = imageRepository.findImageById(imageId);
+export async function getImage(imageId) {
+  if (imageId === null) {
+    return Result.fail(AppError.validationError("Invalid ImageId"));
+  }
+  const result = await imageRepository.findById(imageId);
   return result;
 }
 
-//画像のパス取得
-export async function getImageFilePath(imageId, dirPath) {
-  const result = imageRepository.findByParams(imageId, dirPath);
-  return result;
-}
+//画像の削除
+export async function deleteImage(imageId) {
+  if (imageId === null) {
+    return Result.fail(AppError.validationError("Invalid ImageId"));
+  }
+  const foundPath = await imageRepository.findById(imageId);
+  if (foundPath.isFailure) return foundPath;
+  const uploadPath = foundPath.data;
 
-//画像の削除(レコード)
-export function deleteImageRecord(imageId) {
-  const result = imageRepository.deleteRecord(imageId);
-  return result;
-}
+  const deleteRecordResult = await imageRepository.deleteRecord(imageId);
+  if (deleteRecordResult.isFailure) return deleteRecordResult;
 
-//画像の削除(ファイル)
-export function deleteImageFile(dirPath) {
-  const result = imageRepository.deleteFile(dirPath);
-  return result;
+  const deleteFileResult = await imageRepository.deleteFile(uploadPath);
+  if (deleteFileResult.isFailure) return deleteFileResult;
+
+  return deleteFileResult;
 }
